+++ src/main/kotlin/com/wafflestudio/csereal/common/config/SecurityConfig.kt
@@ -1,6 +1,5 @@
 package com.wafflestudio.csereal.common.config
 
-import com.wafflestudio.csereal.common.mockauth.DevAuthenticationProvider
 import com.wafflestudio.csereal.common.properties.EndpointProperties
 import com.wafflestudio.csereal.core.user.service.CustomOidcUserService
 import jakarta.servlet.http.HttpServletRequest
@@ -10,16 +9,12 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Profile
-import org.springframework.security.authentication.AuthenticationManager
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
 import org.springframework.security.core.Authentication
 import org.springframework.security.web.SecurityFilterChain
 import org.springframework.security.web.authentication.logout.LogoutSuccessHandler
 import org.springframework.security.web.authentication.logout.SimpleUrlLogoutSuccessHandler
-import org.springframework.security.web.context.HttpSessionSecurityContextRepository
-import org.springframework.security.web.context.SecurityContextRepository
 import org.springframework.web.cors.CorsConfiguration
 import org.springframework.web.cors.CorsConfigurationSource
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource
@@ -32,19 +27,8 @@ class SecurityConfig(
     private val customOidcUserService: CustomOidcUserService,
     private val endpointProperties: EndpointProperties,
     @Value("\${login-page}")
-    private val loginPage: String,
-    private val devAuthenticationProvider: DevAuthenticationProvider
+    private val loginPage: String
 ) {
-    @Bean
-    fun authenticationManager(http: HttpSecurity): AuthenticationManager {
-        http.authenticationProvider(devAuthenticationProvider)
-        return http.getSharedObject(AuthenticationManagerBuilder::class.java).build()
-    }
-
-    @Bean
-    fun securityContextRepository(): SecurityContextRepository {
-        return HttpSessionSecurityContextRepository()
-    }
 
     @Bean
     fun filterChain(http: HttpSecurity): SecurityFilterChain {
